fix: sqlx type mismatch (i64 -> i32)

+++ shared/models/src/lib.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone, Hash)]
 pub struct Resource {
-    pub id: i64,
+    pub id: i32,
     pub title: String,
     pub description: String,
     pub year_published: Option<i32>,
@@ -11,13 +11,13 @@ pub struct Resource {
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone, Hash)]
 pub struct Tag {
-    pub id: i64,
+    pub id: i32,
     pub name: String,
 }
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone, Hash)]
 pub struct Tagging {
-    pub id: i64,
-    pub tag_id: i64,
-    pub resource_id: i64,
+    pub id: i32,
+    pub tag_id: i32,
+    pub resource_id: i32,
 }
